Move fakes to pkg/runit
chpst, sv and servicebuilder are runit-related binaries. Presenting fake
versions of them should also be a pkg/runit responsibility.
