--- pkg/hoist/erring_sv
@@ -1,4 +1,0 @@
-#!/usr/bin/env sh
-
-echo $@ >&2
-exit 1
