--- pkg/hoist/fake_servicebuilder
@@ -1,3 +1,0 @@
-#!/usr/bin/env sh
-
-echo $@
