--- pkg/hoist/fake_sv
@@ -1,3 +1,0 @@
-#!/usr/bin/env sh
-
-echo $@
