+++ pkg/runit/runit_service_test.go
@@ -1,15 +1,13 @@
 package runit
 
 import (
-	"runtime"
 	"testing"
 
 	. "github.com/anthonybishopric/gotcha"
-	"github.com/square/p2/pkg/util"
 )
 
 func TestRunitServicesCanBeStarted(t *testing.T) {
-	sv := SV{util.From(runtime.Caller(0)).ExpandPath("fake_sv")}
+	sv := FakeSV()
 	service := &Service{"/var/service/foo", "foo"}
 	out, err := sv.Start(service)
 	Assert(t).IsNil(err, "There should not have been an error starting the service")
@@ -17,7 +15,7 @@ func TestRunitServicesCanBeStarted(t *testing.T) {
 }
 
 func TestErrorReturnedIfRunitServiceBails(t *testing.T) {
-	sv := SV{util.From(runtime.Caller(0)).ExpandPath("erring_sv")}
+	sv := ErringSV()
 	service := &Service{"/var/service/foo", "foo"}
 	_, err := sv.Start(service)
 	Assert(t).IsNotNil(err, "There should have been an error starting the service")
